+++ src/app/admin/admin.module.ts
@@ -41,9 +41,9 @@ import { AdminBlocks } from "./components/adminBlocks/adminBlocks.component";
 import { AdminDashboard } from "./components/adminDashboard/adminDashboard.component";
 import { AdminFilters } from "./components/adminFilters/adminFilters.component";
 import { AdminReports } from "./components/adminReports/adminReports.component";
-import { PostEditForm } from "@app/components/postEditForm/postEditForm.component";
-import { ItemDeleteForm } from "@app/components/itemDeleteForm/itemDeleteForm.component";
-import { DisplayNameEditForm } from "@app/components/displayNameEditForm/displayNameEditForm.component";
+import { PostEditForm } from "@app/components/forms/postEditForm/postEditForm.component";
+import { ItemDeleteForm } from "@app/components/forms/itemDeleteForm/itemDeleteForm.component";
+import { DisplayNameEditForm } from "@app/components/forms/displayNameEditForm/displayNameEditForm.component";
 import { Loader } from "@app/components/loader/loader.component";
 
 @NgModule({
