+++ src/app/components/myPosts/myPosts.component.ts
@@ -45,7 +45,7 @@ import { SWManager } from "@app/services/sWManager.service";
 import { ApiClientService } from "@app/services/apiClient.service";
 import { Loader } from "@app/components/loader/loader.component";
 import { SinglePost } from "@app/components/post/post.component";
-import { ItemDeleteForm } from "@app/components/itemDeleteForm/itemDeleteForm.component";
+import { ItemDeleteForm } from "@app/components/forms/itemDeleteForm/itemDeleteForm.component";
 
 interface MyPostsResponse {
   page: number;
