+++ src/app/components/post/post.component.spec.ts
@@ -48,10 +48,10 @@ import { BehaviorSubject } from "rxjs";
 
 import { SinglePost } from "./post.component";
 import { mockAuthedUser } from "@tests/mockData";
-import { ItemDeleteForm } from "@app/components/itemDeleteForm/itemDeleteForm.component";
-import { ReportForm } from "@app/components/reportForm/reportForm.component";
-import { PostEditForm } from "@app/components/postEditForm/postEditForm.component";
-import { SendHugForm } from "@app/components/sendHugForm/sendHugForm.component";
+import { ItemDeleteForm } from "@app/components/forms/itemDeleteForm/itemDeleteForm.component";
+import { ReportForm } from "@app/components/forms/reportForm/reportForm.component";
+import { PostEditForm } from "@app/components/forms/postEditForm/postEditForm.component";
+import { SendHugForm } from "@app/components/forms/sendHugForm/sendHugForm.component";
 import { ItemsService } from "@app/services/items.service";
 import { AuthService } from "@app/services/auth.service";
