+++ src/app/routes/userPage/userPage.component.spec.ts
@@ -49,8 +49,8 @@ import { AuthService } from "@app/services/auth.service";
 import { mockAuthedUser } from "@tests/mockData";
 import { OtherUser } from "@app/interfaces/otherUser.interface";
 import { iconCharacters } from "@app/interfaces/types";
-import { DisplayNameEditForm } from "@app/components/displayNameEditForm/displayNameEditForm.component";
-import { ReportForm } from "@app/components/reportForm/reportForm.component";
+import { DisplayNameEditForm } from "@app/components/forms/displayNameEditForm/displayNameEditForm.component";
+import { ReportForm } from "@app/components/forms/reportForm/reportForm.component";
 import { ApiClientService } from "@app/services/apiClient.service";
 import { Loader } from "@app/components/loader/loader.component";
 import { UserIcon } from "@app/components/userIcon/userIcon.component";
